+++ Changelog.md
@@ -2,6 +2,11 @@
 
 ## 0.1.*
 
+### 0.1.4
+ - + Added `after` rule property
+ - + Added `command` rule property
+ - + Added test for `after` and `command` properties in Rule
+
 ### 0.1.3
  - * Fixed missing Sticker class in exports
  - * Fixed response ID by adding `direct` parameter